Fix msvc warnings.

+++ trunk/src/nvtt/tools/resize.cpp
@@ -84,14 +84,14 @@ int main(int argc, char *argv[])
 		if (strcmp("-s", argv[i]) == 0)
 		{
 			if (i+1 < argc && argv[i+1][0] != '-') {
-				scale = atof(argv[i+1]);
+				scale = (float)atof(argv[i+1]);
 				i++;
 			}
 		}
 		else if (strcmp("-g", argv[i]) == 0)
 		{
 			if (i+1 < argc && argv[i+1][0] != '-') {
-				gamma = atof(argv[i+1]);
+				gamma = (float)atof(argv[i+1]);
 				i++;
 			}
 		}
@@ -155,7 +155,7 @@ int main(int argc, char *argv[])
 	nv::FloatImage fimage(&image);
 	fimage.toLinear(0, 3, gamma);
 	
-	nv::AutoPtr<nv::FloatImage> fresult(fimage.downSample(*filter, image.width() * scale, image.height() * scale, nv::FloatImage::WrapMode_Mirror));
+	nv::AutoPtr<nv::FloatImage> fresult(fimage.downSample(*filter, uint(image.width() * scale), uint(image.height() * scale), nv::FloatImage::WrapMode_Mirror));
 	
 	nv::AutoPtr<nv::Image> result(fresult->createImageGammaCorrect(gamma));
